+++ subcommand/acl-controller/command_ent_test.go
@@ -36,3 +36,16 @@ func TestUpsertConsulResourcesEnt(t *testing.T) {
 		},
 	})
 }
+
+func TestUpsertAnonymousTokenPolicyEnt(t *testing.T) {
+	testUpsertAnonymousTokenPolicy(t, map[string]anonTokenTest{
+		"primary datacenter": {
+			agentConfig: AgentConfig{
+				Config:      Config{Datacenter: "dc1"},
+				DebugConfig: Config{PrimaryDatacenter: "dc1"},
+			},
+			partitionsEnabled: true,
+			expPolicy:         expEntAnonTokenPolicy,
+		},
+	})
+}