+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## Unreleased
+
+BUG FIXES
+* Fix issue where the `acl-controller` did not update the default namespace with the cross-namespace policy.
+  [[GH-104](https://github.com/hashicorp/consul-ecs/pull/104)]
+* Fix token cleanup in the `acl-controller` when Consul Enterprise admin partitions are enabled.
+  [[GH-105](https://github.com/hashicorp/consul-ecs/pull/105)]
+* The `acl-controller` configures the anonymous token with `service:read` and `node:read`
+  permissions to support cross-dc or cross-partition traffic through mesh gateways.
+  [[GH-103](https://github.com/hashicorp/consul-ecs/pull/103)]
+  [[GH-106](https://github.com/hashicorp/consul-ecs/pull/106)]
+
 ## 0.5.0-beta1 (Jun 06, 2022)
 
 BREAKING CHANGES